pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hydra-auto-resume"
version = "0.2.0"
description = "A Hydra plugin for unified checkpoint resumption with WandB and Slurm."
readme = "README.md"
authors = [
  { name = "Yesung Hwang", email = "yesung.hw@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]
dependencies = [
    "hydra-core",
    "wandb",
    "submitit", 
    "omegaconf"
]
requires-python = ">=3.8"

[project.urls]
"Homepage" = "https://github.com/yesung31/hydra-auto-resume"
